refactor(chat): group modules in package src/askui/chat
- rm unused deps
- only gitignore /chat in root
