+++ .gitignore
@@ -163,4 +163,5 @@ cython_debug/
 
 reports/
 .DS_Store
-chat/
+/chat
+
