+++ pyproject.toml
@@ -39,7 +39,7 @@ format = "black ."
 lint = "ruff check ."
 checks = {composite = ["sort . -c", "format . --check"]}
 fix = {composite = ["sort", "format", "lint"]}
-chat = "streamlit run src/askui/chat.py"
+chat = "streamlit run src/askui/chat/__main__.py"
 
 [dependency-groups]
 test = [
@@ -50,7 +50,4 @@ test = [
 ]
 chat = [
     "streamlit>=1.42.0",
-]
-dev = [
-    "watchdog>=6.0.0",
 ]
